% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{assay_equal}
\alias{assay_equal}
\title{Retrieve assays by exact match}
\usage{
assay_equal(
  word = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{word}{A character string of an assay name}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame of assay information for the given input.
}
\description{
Retrieve assays by exact match
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve assays that match with the character string `ATG_STAT3_CIS`
atg_stat3_cis_assay <- assay_equal(word = 'ATG_STAT3_CIS')
atg_stat3_cis_assay
\dontshow{\}) # examplesIf}
}
