% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{assay_equal_batch}
\alias{assay_equal_batch}
\title{Retrieve assays by exact match batch}
\usage{
assay_equal_batch(
  word_list = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{word_list}{A vector of character strings of an assay names}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A named list of data.frames of assay information for the given input.
}
\description{
Retrieve assays by exact match batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve assays that start with the character string `ATG_STAT3_CIS` and
# `APR_HepG2_MicrotubuleCSK_1hr`
assays <- assay_starts_with_batch(word_list = c('ATG_STAT3_CIS',
                                                'APR_HepG2_MicrotubuleCSK_1hr'))
assays
\dontshow{\}) # examplesIf}
}
