% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{assay_starts_with}
\alias{assay_starts_with}
\title{Retrieve assays by starting characters}
\usage{
assay_starts_with(
  word = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE,
  top = NULL
)
}
\arguments{
\item{word}{A character string of an assay name}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}

\item{top}{Limit the number of returned entries.}
}
\value{
A data.frame of assay information for the given input.
}
\description{
Retrieve assays by starting characters
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve assays that start with the character string `ATG_S`
atg_s_assays <- assay_starts_with(word = 'ATG_S')
atg_s_assays
\dontshow{\}) # examplesIf}
}
