% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioactivity_api_server}
\alias{bioactivity_api_server}
\title{Bioactivity API Server url}
\format{
An object of class \code{character} of length 1.
}
\usage{
bioactivity_api_server
}
\description{
A section of url used in Bioactivity API Endpoints
}
\keyword{datasets}
