% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_ctx.R
\name{check_api_key}
\alias{check_api_key}
\title{Check API key}
\usage{
check_api_key(API_key = NULL, verbose = FALSE)
}
\arguments{
\item{API_key}{User input API key during function call.}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
Either the API key (input during function call or loaded) or NULL
(neither input during function call nor loaded).
}
\description{
Check API key
}
