% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemical_api_server}
\alias{chemical_api_server}
\title{Chemical API Server url}
\format{
An object of class \code{character} of length 1.
}
\usage{
chemical_api_server
}
\description{
A section of url used in Chemical API Endpoints
}
\keyword{datasets}
