% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{create_data.table_chemical_details}
\alias{create_data.table_chemical_details}
\title{Create chemical details data.table helper function}
\usage{
create_data.table_chemical_details()
}
\value{
An empty data.table with columns matching the expected format of the
  get_chemical_details API call.
}
\description{
Create chemical details data.table helper function
}
