% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure_api_server}
\alias{exposure_api_server}
\title{Exposure API Server url}
\format{
An object of class \code{character} of length 1.
}
\usage{
exposure_api_server
}
\description{
A section of url used in Exposure API Endpoints
}
\keyword{datasets}
