% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_aed_data}
\alias{get_aed_data}
\title{Get administered equivalent dose (AED) data for a given chemical}
\usage{
get_aed_data(
  DTXSID = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A data.frame of AED data derived from ToxCast in virto bioactivity
data for given DTXSID.
}
\description{
Get administered equivalent dose (AED) data for a given chemical
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get data for DTXSID5021209
aed <- get_aed_data(DTXSID = 'DTXSID5021209')
aed
\dontshow{\}) # examplesIf}
}
