% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_aed_data_batch}
\alias{get_aed_data_batch}
\title{Get administered equivalent dose (AED) data for a given chemical batch}
\usage{
get_aed_data_batch(
  DTXSID = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A data.table of AED data derived from ToxCast in virto bioactivity
data for given DTXSIDs.
}
\description{
Get administered equivalent dose (AED) data for a given chemical batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get data for DTXSID5021209 and DTXSID7020182
aed <- get_aed_data_batch(DTXSID = c('DTXSID5021209', 'DTXSID7020182'))
aed
\dontshow{\}) # examplesIf}
}
