% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_aeid_by_endpoint}
\alias{get_aeid_by_endpoint}
\title{Get AEID by assay component endpoint name}
\usage{
get_aeid_by_endpoint(
  endpoint = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{The assay component endpoint name.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
An integer corresponding to the AEID of the given assay endpoint.
}
\description{
Get AEID by assay component endpoint name
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve AEID for end point 'APR_HepG2_MicrotubuleCSK_1hr'
aeid <- get_aeid_by_endpoint(endpoint = 'APR_HepG2_MicrotubuleCSK_1hr')
aeid
\dontshow{\}) # examplesIf}
}
