% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_all_assays}
\alias{get_all_assays}
\title{Retrieve all assays}
\usage{
get_all_assays(
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame containing all the assays and associated information
}
\description{
Retrieve all assays
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve all assays
assays <- get_all_assays()
\dontshow{\}) # examplesIf}
}
