% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_analytical_qc}
\alias{get_analytical_qc}
\title{Get analytical QC data for a chemical}
\usage{
get_analytical_qc(
  DTXSID = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A data.frame of analytical QC data for the requested DTXSID.
}
\description{
Get analytical QC data for a chemical
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get QC data for DTXSID7020182
bpa_aq <- get_analytical_qc(DTXSID = 'DTXSID7020182')
bpa_aq
\dontshow{\}) # examplesIf}
}
