% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_analytical_qc_batch}
\alias{get_analytical_qc_batch}
\title{Get analytical QC data for a chemical batch}
\usage{
get_analytical_qc_batch(
  DTXSID = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A named list of data.frames of analytical QC data for the requested
DTXSIDs.
}
\description{
Get analytical QC data for a chemical batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get QC data for DTXSID7020182 and DTXSID0020232
aq <- get_analytical_qc(DTXSID = c('DTXSID7020182', 'DTXSID0020232'))
aq
\dontshow{\}) # examplesIf}
}
