% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_aop_data_batch}
\alias{get_aop_data_batch}
\title{Get ToxCast-mapped AOP data batch}
\usage{
get_aop_data_batch(
  AEID = NULL,
  KeyEvent = NULL,
  EntrezGeneId = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEIDs}

\item{KeyEvent}{The Key Event numbers}

\item{EntrezGeneId}{The Entrez Gene IDs}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A named list of data.frames of ToxCast-mapped AOP data for the
given input.
}
\description{
Get ToxCast-mapped AOP data batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# By AEID, Key Event, and Entrez Gene ID
aop_entrez <- get_aop_data_batch(EntrezGeneId = 196)
aop_entrez
aop_ke <- get_aop_data_batch(KeyEvent = 18)
aop_ke
aop_aeid <- get_aop_data_batch(AEID = 63)
aop_aeid
\dontshow{\}) # examplesIf}
}
