% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_assay_summary_by_gene}
\alias{get_assay_summary_by_gene}
\title{Get assay summary data by gene symbol}
\usage{
get_assay_summary_by_gene(
  geneSymbol = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{geneSymbol}{The gene symbol.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame of assay summary data for applicable assays for
requested offical gene symbol.
}
\description{
Get assay summary data by gene symbol
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve summary data for gene symbol TUBA1A
summary_tuba1a <- get_assay_summary_by_gene(geneSymbol = 'TUBA1A')
summary_tuba1a
\dontshow{\}) # examplesIf}
}
