% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_assay_summary_by_gene_batch}
\alias{get_assay_summary_by_gene_batch}
\title{Get assay summary data by gene symbol batch}
\usage{
get_assay_summary_by_gene_batch(
  geneSymbol = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{geneSymbol}{The gene symbol.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A named list of data.frames of assay summary data for applicable
assays for requested offical gene symbols.
}
\description{
Get assay summary data by gene symbol batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve summary data for gene symbol TUBA1A and IYD
summary <- get_assay_summary_by_gene_batch(geneSymbol = c('TUBA1A', 'IYD'))
summary
\dontshow{\}) # examplesIf}
}
