% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_bioactivity_summary_by_tissue}
\alias{get_bioactivity_summary_by_tissue}
\title{Get summary data by DTXSID and assay tissue origin}
\usage{
get_bioactivity_summary_by_tissue(
  DTXSID = NULL,
  Tissue = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{Tissue}{The tissue of origin for the assay}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame of summary data for the given chemical and tissue.
}
\description{
Get summary data by DTXSID and assay tissue origin
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get data for DTXSID7020192 and liver
liver_bpa <- get_bioactivity_summary_by_tissue(DTXSID = 'DTXSID7020182',
                                               Tissue = 'liver')
liver_bpa
\dontshow{\}) # examplesIf}
}
