% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_bioactivity_summary_by_tissue_batch}
\alias{get_bioactivity_summary_by_tissue_batch}
\title{Get summary data by DTXSID and assay tissue origin batch}
\usage{
get_bioactivity_summary_by_tissue_batch(
  DTXSID = NULL,
  Tissue = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{Tissue}{The tissue of origin for the assay}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A named list of data.frames of summary data for the given chemicals and tissue.
}
\description{
Get summary data by DTXSID and assay tissue origin batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get data for DTXSID7020182 and DTXSID7024241 and liver
liver_summary <- get_bioactivity_summary_by_tissue(DTXSID = c('DTXSID7020182', 'DTXSID7024241'),
                                                   Tissue = 'liver')
liver_summary
\dontshow{\}) # examplesIf}
}
