% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_biomonitoring_data}
\alias{get_biomonitoring_data}
\title{Get Biomonitoring data}
\usage{
get_biomonitoring_data(
  DTXSID = NULL,
  API_key = NULL,
  Projection = "",
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Projection}{Optional parameter controlling return type.}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame containing general use keywords.
}
\description{
Get Biomonitoring data
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get biomonitoring data for Caffeine
get_biomonitoring_data('DTXSID0020232')
\dontshow{\}) # examplesIf}
}
