% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_cancer_hazard_batch_old}
\alias{get_cancer_hazard_batch_old}
\title{Get cancer hazard batch}
\usage{
get_cancer_hazard_batch_old(DTXSID = NULL, API_key = NULL, rate_limit = 0L)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between requests}
}
\value{
A named list of data.frames, each containing cancer hazard and
  related data for each input DTXSID.
}
\description{
Get cancer hazard batch
}
