% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_chem_info_batch_old}
\alias{get_chem_info_batch_old}
\title{Retrieve chemical information in batch search}
\usage{
get_chem_info_batch_old(
  DTXSID = NULL,
  type = "",
  API_key = NULL,
  rate_limit = 0L
)
}
\arguments{
\item{DTXSID}{A vector of chemical identifier DTXSIDs}

\item{type}{A vector of type used in get_chem_info(). This specifies whether
to only grab predicted or experimental results. If not specified, it will
grab all details. The allowable input values are "", predicted", or
"experimental".}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between each request}
}
\value{
A named list of data.frames containing chemical information for the
  chemicals with DTXSID matching the input parameter.
}
\description{
Retrieve chemical information in batch search
}
