% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_chem_props_exp_batch}
\alias{get_chem_props_exp_batch}
\title{Get experimental physical-chemical property data via batch}
\usage{
get_chem_props_exp_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = chemical_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between each request}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.table of experimental physchem property data
}
\description{
Get experimental physical-chemical property data via batch
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Get experimental physchem properties for BPA and Caffeine
chem_props_exp <- get_chem_props_exp(DTXSID = c('DTXSID7020182',
                                                'DTXSID0020232'))
\dontshow{\}) # examplesIf}
}
