% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_details_dtxsid}
\alias{get_chemical_details_dtxsid}
\title{Retrieve chemical details from DTXSID}
\usage{
get_chemical_details_dtxsid(DTXSID, API_key)
}
\arguments{
\item{DTXSID}{The chemical identifer DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame containing chemical information for the chemical with
  DTXSID matching the input parameter.
}
\description{
Retrieve chemical details from DTXSID
}
