% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_mrv_by_dtxsid}
\alias{get_chemical_mrv_by_dtxsid}
\title{Get mrv file by DTXSID}
\usage{
get_chemical_mrv_by_dtxsid(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifer DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
XML file format for representing a mrv file.
}
\description{
Get mrv file by DTXSID
}
