% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_chemicals_by_assay}
\alias{get_chemicals_by_assay}
\title{Get list of chemical DTXSIDs for a given assay}
\usage{
get_chemicals_by_assay(
  AEID = NULL,
  Projection = "dtxsidonly",
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEID}

\item{Projection}{The format and DTXSID data returned. Allowed values are
'dtxsidonly' and 'ccdassaydetails'. The default format
is 'dtxsidonly'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A list of DTXSIDs or data.frame of assay information.
}
\description{
Get list of chemical DTXSIDs for a given assay
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
dtxsid_list <- get_chemicals_by_assay(AEID = 3032)
\dontshow{\}) # examplesIf}
}
