% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_chemicals_by_assay_batch}
\alias{get_chemicals_by_assay_batch}
\title{Get list of chemical DTXSIDs for a given assay batch}
\usage{
get_chemicals_by_assay_batch(
  AEID = NULL,
  Projection = "dtxsidonly",
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEIDs}

\item{Projection}{The format and DTXSID data returned. Allowed values are
'dtxsidonly' and 'ccdassaydetails'. The default format
is 'dtxsidonly'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A named list of DTXSIDs or data.frame of assay information.
}
\description{
Get list of chemical DTXSIDs for a given assay batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
dtxsid_list <- get_chemicals_by_assay_batch(AEID = c(3032, 743))
dtxsid_list
\dontshow{\}) # examplesIf}
}
