% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs-batch.R
\name{get_demographic_exposure_prediction_batch}
\alias{get_demographic_exposure_prediction_batch}
\title{Retrieve demographic exposure predictions for chemicals via batch}
\usage{
get_demographic_exposure_prediction_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = exposure_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between each request}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A named list of data.frames, each containing demographic exposure
prediction data for each input DTXSID.
}
\description{
Retrieve demographic exposure predictions for chemicals via batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Pull demographic exposure prediction data for multiple chemicals
dtxsid <- c('DTXSID7020182', 'DTXSID2021315')
exp_demo <- get_demographic_exposure_prediction_batch(DTXSID = dtxsid)
\dontshow{\}) # examplesIf}
}
