% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_exposure_endpoint_status}
\alias{get_exposure_endpoint_status}
\title{Exposure API Endpoint status}
\usage{
get_exposure_endpoint_status()
}
\value{
Status of Exposure API Endpoints
}
\description{
Exposure API Endpoint status
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
status <- get_exposure_endpoint_status()
print(status)
\dontshow{\}) # examplesIf}
}
