% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs-batch.R
\name{get_exposure_list_presence_tags_batch}
\alias{get_exposure_list_presence_tags_batch}
\title{Retrieve document data and list presence tags for chemicals batch}
\usage{
get_exposure_list_presence_tags_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = exposure_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between each request}

\item{Server}{The root address for the API endpoint}

\item{Verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A named list of data.frames, each containing exposure list presence
tags use data for each input DTXSID.
}
\description{
Retrieve document data and list presence tags for chemicals batch
}
\examples{
has_ccte_key() & is.na(ccte_key() == 'FAKE_KEY')
# Pull exposure functional use data for multiple chemicals
dtxsid <- c('DTXSID7020182', 'DTXSID2021315')
exp_list_tags <- get_exposure_list_presence_tags_batch(DTXSID = dtxsid)
}
