% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_human_hazard_by_dtxsid_batch_old}
\alias{get_human_hazard_by_dtxsid_batch_old}
\title{Get human hazard data by DTXSID batch}
\usage{
get_human_hazard_by_dtxsid_batch_old(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L
)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs.}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between each request}
}
\value{
A named lit of data.frames containing chemical human hazard data.
}
\description{
Get human hazard data by DTXSID batch
}
