% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_ms_ready_by_formula_batch}
\alias{get_ms_ready_by_formula_batch}
\title{Get msready by formula batch search}
\usage{
get_ms_ready_by_formula_batch(formula_list = NULL, API_key = NULL)
}
\arguments{
\item{formula_list}{A list of strings denoting the input chemicals formulas}

\item{API_key}{The user-specific API key}
}
\value{
A named list of character lists of DTXSIDs with chemical formulas
  matching the search criteria
}
\description{
Get msready by formula batch search
}
