% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_ms_ready_by_mass_batch}
\alias{get_ms_ready_by_mass_batch}
\title{Get ms ready by mass batch search}
\usage{
get_ms_ready_by_mass_batch(start_list = NULL, end_list = NULL, API_key = NULL)
}
\arguments{
\item{start_list}{A numeric list of starting values for mass range}

\item{end_list}{A numeric list of ending values for mass range}

\item{API_key}{The user-specific API key}
}
\value{
A named list of character lists with DTXSIDs with msready masses
  falling within the given ranges.
}
\description{
Get ms ready by mass batch search
}
