% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_predictions_by_dtxsid}
\alias{get_predictions_by_dtxsid}
\title{Get bioactivity model predictions by DTXSID}
\usage{
get_predictions_by_dtxsid(
  DTXSID = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A data.frame of ToxCast model prediction data for given DTXSID.
}
\description{
Get bioactivity model predictions by DTXSID
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get predictions for DTXSID70201082
bpa_predictions <- get_predictions_by_dtxsid(DTXSID = 'DTXSID7020182')
bpa_predictions
\dontshow{\}) # examplesIf}
}
