% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_predictions_by_dtxsid_and_model}
\alias{get_predictions_by_dtxsid_and_model}
\title{Get bioactivity model predictions by DTXSID and Model}
\usage{
get_predictions_by_dtxsid_and_model(
  DTXSID = NULL,
  Model = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{Model}{The ToxCast model type. Model type options include: 'CERAPP
Potency Level (Consensus)', 'CERAPP Potency Level (From Literature)',
'COMPARA (Consensus)', and 'ToxCast Pathway Model (AUC)'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A data.frame of ToxCast model predictions for the given DTXSID and
Model.
}
\description{
Get bioactivity model predictions by DTXSID and Model
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get predictions for DTXSID70201082 and CERAPP models
bpa_predictions <- get_predictions_by_dtxsid_and_model(DTXSID = 'DTXSID7020182',
                                                       Model = 'CERAPP')
bpa_predictions
\dontshow{\}) # examplesIf}
}
