% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_predictions_by_dtxsid_and_model_batch}
\alias{get_predictions_by_dtxsid_and_model_batch}
\title{Get bioactivity model predictions by DTXSID and Model batch}
\usage{
get_predictions_by_dtxsid_and_model_batch(
  DTXSID = NULL,
  Model = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{Model}{The ToxCast model type. Model type options include: 'CERAPP
Potency Level (Consensus)', 'CERAPP Potency Level (From Literature)',
'COMPARA (Consensus)', and 'ToxCast Pathway Model (AUC)'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
A named list of data.frames of ToxCast model prediction data
for given DTXSIDs and Model.
}
\description{
Get bioactivity model predictions by DTXSID and Model batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get predictions for DTXSID70201082 and DTXSID0020232 and Model CERAPP
predictions <- get_predictions_by_dtxsid_batch(DTXSID = c('DTXSID7020182',
                                                          'DTXSID0020232'),
                                               Model = 'CERAPP')
predictions
\dontshow{\}) # examplesIf}
}
