% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs-batch.R
\name{get_reported_functional_use_batch}
\alias{get_reported_functional_use_batch}
\title{Get reported functional use via batch}
\usage{
get_reported_functional_use_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between requests}

\item{Server}{The root address for the API endpoint.}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A list of data.frames of reported functional use corresponding to
the input DTXSIDs.
}
\description{
Get reported functional use via batch
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Retrieve reported functional use for BPA and Caffeine
get_reported_functional_use_batch(DTXSID = c('DTXSID7020182',
                                             'DTXSID0020232'))
\dontshow{\}) # examplesIf}
}
