% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_single_concentration}
\alias{get_single_concentration}
\title{Get single concentration data}
\usage{
get_single_concentration(
  AEID = NULL,
  Projection = "single-conc",
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEID}

\item{Projection}{The format and concentration data returned. Allowed values
are 'single-conc' and 'ccd-single-conc'. The default format is
'single-conc'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A data.frame of single concentration screening data for requested
ToxCast assay component endpoint ID (AEID).
}
\description{
Get single concentration data
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get single conc data for AEID 3032 in different projections
aeid_3032 <- get_single_concentration(AEID = 3032)
aeid_ccd_3032 <- get_single_concentration(AEID = 3032,
                                          Projection = 'ccd-single-conc')
\dontshow{\}) # examplesIf}
}
