% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_single_concentration_batch}
\alias{get_single_concentration_batch}
\title{Get single concentration data batch}
\usage{
get_single_concentration_batch(
  AEID = NULL,
  Projection = "single-conc",
  API_key = NULL,
  Server = bioactivity_api_server,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEIDs}

\item{Projection}{The format and concentration data returned. Allowed values
are 'single-conc' and 'ccd-single-conc'. The default format is
'single-conc'.}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A named list of data.frames of single concentration screening data
for requested ToxCast assay component endpoint IDs (AEID).
}
\description{
Get single concentration data batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Get single conc data for AEID 3032 and 743
aeid_concs <- get_single_concentration_batch(AEID = c(3032, 743))
\dontshow{\}) # examplesIf}
}
