% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_skin_eye_hazard_batch_old}
\alias{get_skin_eye_hazard_batch_old}
\title{Get skin and eye hazard batch}
\usage{
get_skin_eye_hazard_batch_old(DTXSID = NULL, API_key = NULL, rate_limit = 0L)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between each request}
}
\value{
A named list of data.frames containing skin and eye hazard data for
  each input DTXSID.
}
\description{
Get skin and eye hazard batch
}
