% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_total_assay_count}
\alias{get_total_assay_count}
\title{Get total assay count}
\usage{
get_total_assay_count(
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be
given.}
}
\value{
An integer indicating the total number of assays.
}
\description{
Get total assay count
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
assay_count <- get_total_assay_count()
print(assay_count)
\dontshow{\}) # examplesIf}
}
