% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hazard_api_server}
\alias{hazard_api_server}
\title{Hazard API Server url}
\format{
An object of class \code{character} of length 1.
}
\usage{
hazard_api_server
}
\description{
A section of url used in Hazard API Endpoints
}
\keyword{datasets}
