% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{prepare_word}
\alias{prepare_word}
\title{Prepare url helper function}
\usage{
prepare_word(word)
}
\arguments{
\item{word}{A character string}
}
\value{
A character string that is ready for use in http request
}
\description{
Prepare url helper function
}
