% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{assay_contains}
\alias{assay_contains}
\title{Retrieve assays by substring}
\usage{
assay_contains(
  word = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE,
  top = NULL
)
}
\arguments{
\item{word}{A character string of an assay name}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}

\item{top}{Limit the number of returned entries.}
}
\value{
A data.frame of assay information for the given input.
}
\description{
Retrieve assays by substring
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# Retrieve assays that contain with the character string `AT3_CIS`
at3_cis_assays <- assay_contains(word = 'AT3_CIS')
at3_cis_assays
\dontshow{\}) # examplesIf}
}
