% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{chemical_contains}
\alias{chemical_contains}
\title{Chemical contains}
\usage{
chemical_contains(word = NULL, API_key = NULL)
}
\arguments{
\item{word}{A character string of a chemical name or portion of a chemical
name}

\item{API_key}{The user-specific API key}
}
\value{
A list of DTXSIDs of chemicals matching the query parameters
}
\description{
Chemical contains
}
