% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{chemical_contains_batch}
\alias{chemical_contains_batch}
\title{Chemical contains batch search}
\usage{
chemical_contains_batch(wordlist = NULL, API_key = NULL)
}
\arguments{
\item{wordlist}{A list of character strings of chemical names or portion of
chemical names}

\item{API_key}{User-specific API key}
}
\value{
A named list of data.frames of chemicals and related values matching
  the query parameters
}
\description{
Chemical contains batch search
}
