% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{chemical_equal}
\alias{chemical_equal}
\title{Chemical equal}
\usage{
chemical_equal(word = NULL, API_key = NULL)
}
\arguments{
\item{word}{A character string of a chemical name or portion of a chemical
name}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame of chemicals and related values matching the query
  parameters
}
\description{
Chemical equal
}
