% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{chemical_starts_with}
\alias{chemical_starts_with}
\title{Chemical starts with}
\usage{
chemical_starts_with(
  word = NULL,
  API_key = NULL,
  Server = chemical_api_server,
  verbose = FALSE
)
}
\arguments{
\item{word}{A character string of a chemical name or portion of a chemical
name}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame of chemicals and related values matching the query
  parameters
}
\description{
Chemical starts with
}
