% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{chemical_starts_with_batch}
\alias{chemical_starts_with_batch}
\title{Chemical starts with batch search}
\usage{
chemical_starts_with_batch(word_list = NULL, API_key = NULL)
}
\arguments{
\item{word_list}{A list of character strings of chemical names or portion of
chemical names}

\item{API_key}{User-specific API key}
}
\value{
A named list of data.frames of chemicals and related values matching
  the query parameters
}
\description{
Chemical starts with batch search
}
