% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{create_hazard_data.table}
\alias{create_hazard_data.table}
\title{Hazard data.table helper function}
\usage{
create_hazard_data.table()
}
\value{
A data.table used in various functions that wrap the hazard API
endpoints.
}
\description{
Hazard data.table helper function
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hazard_dt <- create_hazard_data.table()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
