% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_all_public_chemical_lists}
\alias{get_all_public_chemical_lists}
\title{Get all public chemical lists}
\usage{
get_all_public_chemical_lists(API_key = NULL)
}
\arguments{
\item{API_key}{The user-specific api key}
}
\value{
A data.frame containing information on all public chemical lists
  available from the CCTE chemical api.
}
\description{
Get all public chemical lists
}
