% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_aop_data}
\alias{get_aop_data}
\title{Get ToxCast-mapped AOP data}
\usage{
get_aop_data(
  AEID = NULL,
  KeyEvent = NULL,
  EntrezGeneId = NULL,
  API_key = NULL,
  Server = bioactivity_api_server,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{The assay endpoint identifier AEID}

\item{KeyEvent}{The Key Event number}

\item{EntrezGeneId}{The Entrez Gene ID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be}
}
\value{
a data.frame of ToxCast-mapped AOP data for the given input.
}
\description{
Get ToxCast-mapped AOP data
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) withAutoprint(\{ # examplesIf}
# By AEID, Key Event, and Entrez Gene ID
aop_entrez <- get_aop_data(EntrezGeneId = 196)
aop_entrez
aop_ke <- get_aop_data(KeyEvent = 18)
aop_ke
aop_aeid <- get_aop_data(AEID = 63)
aop_aeid
\dontshow{\}) # examplesIf}
}
