% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_bioactivity_details}
\alias{get_bioactivity_details}
\title{Retrieve bioactivity data from DTXSID or AEID}
\usage{
get_bioactivity_details(DTXSID = NULL, AEID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{AEID}{The chemical identifier AEID}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame containing bioactivity information for the chemical with
  DTXSID or AEID matching the input parameter.
}
\description{
Retrieve bioactivity data from DTXSID or AEID
}
