% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_bioactivity_details_batch}
\alias{get_bioactivity_details_batch}
\title{Retrieve bioactivity data from DTXSID or AEID batch}
\usage{
get_bioactivity_details_batch(DTXSID = NULL, AEID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs.}

\item{AEID}{A list of assay endpoint identifiers AEIDs.}

\item{API_key}{The user-specific API key.}
}
\value{
A named list of data.frames containing bioactivity information for the chemicals with
  DTXSID or assays with AEID matching the input parameter.
}
\description{
Retrieve bioactivity data from DTXSID or AEID batch
}
