% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs.R
\name{get_bioactivity_summary}
\alias{get_bioactivity_summary}
\title{Retrieve bioactivity summary for AEID}
\usage{
get_bioactivity_summary(
  AEID = NULL,
  API_key = NULL,
  Server = bioactivity_api_server
)
}
\arguments{
\item{AEID}{The assay endpoint indentifier AEID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.frame containing summary information corresponding to the
  input AEID
}
\description{
Retrieve bioactivity summary for AEID
}
