% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity-APIs-batch.R
\name{get_bioactivity_summary_batch}
\alias{get_bioactivity_summary_batch}
\title{Retrieve bioactivity summary data from AEID batch}
\usage{
get_bioactivity_summary_batch(
  AEID = NULL,
  API_key = NULL,
  Server = NULL,
  rate_limit = 0L,
  verbose = FALSE
)
}
\arguments{
\item{AEID}{A list of AEID identifiers}

\item{API_key}{The user-specific API key.}

\item{Server}{The root address for the API endpoint}

\item{rate_limit}{Number of seconds to wait between each request}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A named list of data.frames containing bioactivity summary
information for the assays with AEID matching the input parameter.
}
\description{
Retrieve bioactivity summary data from AEID batch
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get bioactivity summary for multiple aeids
aeids <- get_bioactivity_summary_batch(AEID = c(159, 160))
\dontshow{\}) # examplesIf}
}
