% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs.R
\name{get_cancer_hazard}
\alias{get_cancer_hazard}
\title{Get cancer hazard}
\usage{
get_cancer_hazard(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame of cancer hazard data related to the input DTXSID.
}
\description{
Get cancer hazard
}
