% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chem_info}
\alias{get_chem_info}
\title{Retrieve chemical information}
\usage{
get_chem_info(
  DTXSID = NULL,
  type = c("", "predicted", "experimental"),
  API_key = NULL
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{type}{This specifies whether to only grab predicted or experimental
results. If not specified, it will grab all details. The allowable input
values are "predicted" or "experimental".}

\item{API_key}{The user-specific API Key}
}
\value{
A data.frame containing chemical information for the chemical with
  DTXSID matching the input parameter.
}
\description{
Retrieve chemical information
}
