% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chem_props_exp}
\alias{get_chem_props_exp}
\title{Get experimental physical-chemical property data}
\usage{
get_chem_props_exp(
  DTXSID = NULL,
  API_key = NULL,
  Server = chemical_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame of experimental physchem property data
}
\description{
Get experimental physical-chemical property data
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get experimental physchem properties for BPA
bpa_exp_props <- get_chem_props_exp(DTXSID = 'DTXSID7020182')
\dontshow{\}) # examplesIf}
}
