% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chem_props_pred}
\alias{get_chem_props_pred}
\title{Get predicted physical-chemical property data}
\usage{
get_chem_props_pred(
  DTXSID = NULL,
  API_key = NULL,
  Server = chemical_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame of predicted physchem property data
}
\description{
Get predicted physical-chemical property data
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Get predicted physchem properties for BPA
bpa_pred_props <- get_chem_props_pred(DTXSID = 'DTXSID7020182')
\dontshow{\}) # examplesIf}
}
