% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chem_props_summary}
\alias{get_chem_props_summary}
\title{Get Summary information on chemical properties}
\usage{
get_chem_props_summary(
  DTXSID = NULL,
  API_key = NULL,
  Server = chemical_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID.}

\item{API_key}{The user-specific API key.}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame of summary data for chemical properties.
}
\description{
Get Summary information on chemical properties
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Get summary data for BPA
bpa_props_summary <- get_chem_props_summary(DTXSID = 'DTXSID7020182')
\dontshow{\}) # examplesIf}
}
