% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_by_property_range}
\alias{get_chemical_by_property_range}
\title{Get chemicals by property and its value range}
\usage{
get_chemical_by_property_range(
  start = NULL,
  end = NULL,
  property = NULL,
  API_key = NULL
)
}
\arguments{
\item{start}{A numeric value, the beginning of the range}

\item{end}{A numeric value, the end of the range}

\item{property}{A string, the property in question}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame containing chemical information for chemicals matching
  the search criteria.
}
\description{
Get chemicals by property and its value range
}
