% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_endpoint_status}
\alias{get_chemical_endpoint_status}
\title{Chemical API Endpoint status}
\usage{
get_chemical_endpoint_status()
}
\value{
Status of Chemical API Endpoints
}
\description{
Chemical API Endpoint status
}
\examples{
\dontshow{if (has_ctx_key() & is.na(ctx_key() == 'FAKE_KEY')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
status <- get_chemical_endpoint_status()
print(status)
\dontshow{\}) # examplesIf}
}
