% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_chemical_image_batch}
\alias{get_chemical_image_batch}
\title{Get image file by DTXSID or DTXCID batch}
\usage{
get_chemical_image_batch(
  DTXSID = NULL,
  DTXCID = NULL,
  format = "",
  API_key = NULL
)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs.}

\item{DTXCID}{A list of chemical identifier DTXCIDs.}

\item{format}{The image type, either "png" or "svg". If left blank, will
default to "png".}

\item{API_key}{The user-specific API key.}
}
\value{
A named list of Large arrays of three dimensions representing an image. For
  displaying an image, one may use \code{png::writePNG()} or
  \code{countcolors::plotArrayAsImage()} among many such functions.
}
\description{
Get image file by DTXSID or DTXCID batch
}
