% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_lists_by_type}
\alias{get_chemical_lists_by_type}
\title{Get chemical lists by type}
\usage{
get_chemical_lists_by_type(type = NULL, API_key = NULL)
}
\arguments{
\item{type}{The type of list. This is a case sensitive parameter and returns
lists only for values of "federal", "international", "state", and "other".}

\item{API_key}{The user-specified API key.}
}
\value{
A data.frame containing information about lists that meet the search
  criteria.
}
\description{
Get chemical lists by type
}
