% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_mol}
\alias{get_chemical_mol}
\title{Get mol file by DTXSID or DTXCID}
\usage{
get_chemical_mol(DTXSID = NULL, DTXCID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{DTXCID}{Chemical identifier DTXCID}

\item{API_key}{The user-specific API key}
}
\value{
A character string giving a mol file representation
}
\description{
Get mol file by DTXSID or DTXCID
}
