% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_mrv}
\alias{get_chemical_mrv}
\title{Get mrv file by DTXSID}
\usage{
get_chemical_mrv(DTXSID = NULL, DTXCID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{DTXCID}{The chemical identifier DTXCID}

\item{API_key}{The user-specific API key}
}
\value{
XML file format for representing a mrv file.
}
\description{
Get mrv file by DTXSID
}
