% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemical_synonym}
\alias{get_chemical_synonym}
\title{Get chemical synonym}
\usage{
get_chemical_synonym(
  DTXSID = NULL,
  API_key = NULL,
  Server = chemical_api_server
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A named list of synonym information for the input DTXSID
}
\description{
Get chemical synonym
}
