% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_chemical_synonym_batch}
\alias{get_chemical_synonym_batch}
\title{Get chemical synonym batch}
\usage{
get_chemical_synonym_batch(DTXSID = NULL, API_key = NULL, rate_limit = 0L)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{The number of seconds to wait between requests.}
}
\value{
A named list of lists containing synonym information for each input
  DTXSID.
}
\description{
Get chemical synonym batch
}
