% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_chemical_weight_fraction}
\alias{get_chemical_weight_fraction}
\title{Get Chemical Weight Fractions}
\usage{
get_chemical_weight_fraction(
  DTXSID = NULL,
  API_key = NULL,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame containing chemical weight fraction data.
}
\description{
Get Chemical Weight Fractions
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Get chemical weight fraction data for Caffeine
get_chemical_weight_fraction('DTXSID0020232')
\dontshow{\}) # examplesIf}
}
