% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_chemicals_in_list}
\alias{get_chemicals_in_list}
\title{Get chemicals in a given chemical list}
\usage{
get_chemicals_in_list(list_name = NULL, API_key = NULL)
}
\arguments{
\item{list_name}{The name of the list of chemicals}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame of the chemical list
}
\description{
Get chemicals in a given chemical list
}
