% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_chemicals_in_list_batch}
\alias{get_chemicals_in_list_batch}
\title{Get chemicals in a given chemical list batch}
\usage{
get_chemicals_in_list_batch(list_names = NULL, API_key = NULL)
}
\arguments{
\item{list_names}{A list of names of chemical lists.}

\item{API_key}{The user-specific API key.}
}
\value{
A named list of data.frames each containing chemicals in the
  corresponding chemical lists.
}
\description{
Get chemicals in a given chemical list batch
}
