% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs.R
\name{get_ecotox_hazard_by_dtxsid}
\alias{get_ecotox_hazard_by_dtxsid}
\title{Get ecotox hazard data by DTXSID}
\usage{
get_ecotox_hazard_by_dtxsid(
  DTXSID = NULL,
  API_key = NULL,
  Server = hazard_api_server
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.frame containing chemical (ecotox) hazard data
}
\description{
Get ecotox hazard data by DTXSID
}
