% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_ecotox_hazard_by_dtxsid_batch}
\alias{get_ecotox_hazard_by_dtxsid_batch}
\title{Get ecotox hazard data by DTXSID batch}
\usage{
get_ecotox_hazard_by_dtxsid_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = hazard_api_server
)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs.}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between each request}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.table containing chemical ecotox hazard data.
}
\description{
Get ecotox hazard data by DTXSID batch
}
