% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_exposure_functional_use}
\alias{get_exposure_functional_use}
\title{Retrieve exposure related functional use data}
\usage{
get_exposure_functional_use(
  DTXSID = NULL,
  API_key = NULL,
  Server = "https://api-ccte-stg.epa.gov/exposure"
)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.frame of functional use data.
}
\description{
Retrieve exposure related functional use data
}
