% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_exposure_functional_use_category}
\alias{get_exposure_functional_use_category}
\title{Retireve functional use categories}
\usage{
get_exposure_functional_use_category(
  API_key = NULL,
  Server = "https://api-ccte-stg.epa.gov/exposure"
)
}
\arguments{
\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.frame of functional use categories.
}
\description{
Retireve functional use categories
}
