% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_exposure_functional_use_category}
\alias{get_exposure_functional_use_category}
\title{Retrieve functional use categories}
\usage{
get_exposure_functional_use_category(
  API_key = NULL,
  Server = "https://api-ccte.epa.gov/exposure",
  verbose = FALSE
)
}
\arguments{
\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame of functional use categories.
}
\description{
Retrieve functional use categories
}
\examples{
\dontshow{if (has_ccte_key() & is.na(ccte_key() == 'FAKE_KEY')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Pull functional use category data for BPA
functional_use_categories <- get_exposure_functional_use_category()
\dontshow{\}) # examplesIf}
}
