% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_exposure_functional_use_probability}
\alias{get_exposure_functional_use_probability}
\title{Retrieve probability of exposure for functional use category}
\usage{
get_exposure_functional_use_probability(
  DTXSID = NULL,
  API_key = NULL,
  Server = "https://api-ccte-stg.epa.gov/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.frame with probabilities corresponding to various routes of
  exposure related to functional use.
}
\description{
Retrieve probability of exposure for functional use category
}
