% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_fate_by_dtxsid}
\alias{get_fate_by_dtxsid}
\title{Get fate by DTXSID}
\usage{
get_fate_by_dtxsid(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
@return A data.frame containing chemical information for the chemical with
  DTXSID matching the input parameter.
}
\description{
Get fate by DTXSID
}
