% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_general_use_keywords}
\alias{get_general_use_keywords}
\title{Get General Use Keywords}
\usage{
get_general_use_keywords(
  DTXSID = NULL,
  API_key = NULL,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame containing general use keywords.
}
\description{
Get General Use Keywords
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Get general use keywords for Caffeine
get_general_use_keywords('DTXSID0020232')
\dontshow{\}) # examplesIf}
}
