% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs-batch.R
\name{get_general_use_keywords_batch}
\alias{get_general_use_keywords_batch}
\title{Get general use keywords via batch}
\usage{
get_general_use_keywords_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{Chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between requests}

\item{Server}{The root address for the API endpoint.}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A list of data.frames of general use keywords corresponding to the
input DTXSIDs.
}
\description{
Get general use keywords via batch
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Retrieve general use keywords for BPA and Caffeine
get_general_use_keywords_batch(DTXSID = c('DTXSID7020182', 'DTXSID0020232'))
\dontshow{\}) # examplesIf}
}
