% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs.R
\name{get_genetox_details}
\alias{get_genetox_details}
\title{Get genetox details}
\usage{
get_genetox_details(DTXSID = NULL, API_key = NULL, Server = hazard_api_server)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}
}
\value{
A data.frame of genetox detail data related to the input DTXSID.
}
\description{
Get genetox details
}
