% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_genetox_details_batch}
\alias{get_genetox_details_batch}
\title{Get genetox details batch}
\usage{
get_genetox_details_batch(DTXSID = NULL, API_key = NULL, rate_limit = 0L)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between requests}
}
\value{
A named list of data.frames of genetox detail data for each input
  DTXSID.
}
\description{
Get genetox details batch
}
