% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs.R
\name{get_genetox_summary}
\alias{get_genetox_summary}
\title{Get genetox summary}
\usage{
get_genetox_summary(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame of genetox summary data related to the input DTXSID.
}
\description{
Get genetox summary
}
