% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_genetox_summary_batch}
\alias{get_genetox_summary_batch}
\title{Get genetox summary batch}
\usage{
get_genetox_summary_batch(
  DTXSID = NULL,
  API_key = NULL,
  rate_limit = 0L,
  Server = hazard_api_server,
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key.}

\item{rate_limit}{Number of seconds to wait between requests}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some “progress report” should be given.}
}
\value{
A data.table of genetox summary data for each input
  DTXSID.
}
\description{
Get genetox summary batch
}
