% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_hazard_by_dtxsid_batch}
\alias{get_hazard_by_dtxsid_batch}
\title{Get hazard data by DTXSID batch}
\usage{
get_hazard_by_dtxsid_batch(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs}

\item{API_key}{The user-specific API key}
}
\value{
A named list of data.frames containing chemical (human and eco)
  hazard data for each input chemical.
}
\description{
Get hazard data by DTXSID batch
}
