% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs-batch.R
\name{get_human_hazard_by_dtxsid_batch}
\alias{get_human_hazard_by_dtxsid_batch}
\title{Get human hazard data by DTXSID batch}
\usage{
get_human_hazard_by_dtxsid_batch(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{A list of chemical identifier DTXSIDs.}

\item{API_key}{The user-specific API key.}
}
\value{
A named lit of data.frames containing chemical human hazard data.
}
\description{
Get human hazard data by DTXSID batch
}
