% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_lists_containing_chemical}
\alias{get_lists_containing_chemical}
\title{Get chemical lists containing given chemical}
\usage{
get_lists_containing_chemical(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
A list of names of chemical lists that contain the given chemical
}
\description{
Get chemical lists containing given chemical
}
