% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_medium_categories}
\alias{get_medium_categories}
\title{Retrieve MMDB medium categories}
\usage{
get_medium_categories(
  API_key = NULL,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be
given.}
}
\value{
A data.frame of harmonized medium categories from MMDB and relevant
descriptions.
}
\description{
Retrieve MMDB medium categories
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Retrieve medium categories and descriptions
get_medium_categories()
\dontshow{\}) # examplesIf}
}
