% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_msready_by_dtxcid}
\alias{get_msready_by_dtxcid}
\title{Get msready by DTXCID}
\usage{
get_msready_by_dtxcid(DTXCID = NULL, API_key = NULL)
}
\arguments{
\item{DTXCID}{The chemical identifier DTXCID}

\item{API_key}{The user-specific API key}
}
\value{
A character list of DTXSIDs with DTXCIDs matching the
  search criteria
}
\description{
Get msready by DTXCID
}
