% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_msready_by_formula}
\alias{get_msready_by_formula}
\title{Get msready by formula}
\usage{
get_msready_by_formula(formula = NULL, API_key = NULL)
}
\arguments{
\item{formula}{A string denoting the input chemical formula}

\item{API_key}{The user-specific API key}
}
\value{
A character list of DTXSIDs with chemical formulas matching the
  search criteria
}
\description{
Get msready by formula
}
