% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_msready_by_formula_batch}
\alias{get_msready_by_formula_batch}
\title{Get msready by formula batch search}
\usage{
get_msready_by_formula_batch(
  formula_list = NULL,
  API_key = NULL,
  rate_limit = 0L
)
}
\arguments{
\item{formula_list}{A list of strings denoting the input chemicals formulas}

\item{API_key}{The user-specific API key}

\item{rate_limit}{Number of seconds to wait between each request}
}
\value{
A named list of character lists of DTXSIDs with chemical formulas
  matching the search criteria
}
\description{
Get msready by formula batch search
}
