% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs.R
\name{get_public_chemical_list_by_name}
\alias{get_public_chemical_list_by_name}
\title{Get chemical list by name}
\usage{
get_public_chemical_list_by_name(
  list_name = NULL,
  Projection = "",
  API_key = NULL
)
}
\arguments{
\item{list_name}{The name of the list of chemicals}

\item{Projection}{Optional parameter controlling return type. It takes values
chemicallistall' and 'chemicallistname' with the former as the default
value.}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame containing information about the chemical list. Note,
  this is not the chemical list itself. To access the chemicals in the list,
  use \code{\link{get_chemicals_in_list}}.
}
\description{
Get chemical list by name
}
\seealso{
\code{\link{get_chemicals_in_list}}
}
