% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical-APIs-batch.R
\name{get_public_chemical_list_by_name_batch}
\alias{get_public_chemical_list_by_name_batch}
\title{Get chemical list by name batch}
\usage{
get_public_chemical_list_by_name_batch(name_list = NULL, API_key = NULL)
}
\arguments{
\item{name_list}{A list of chemical list names.}

\item{API_key}{The user-specific API key.}
}
\value{
A named list of data.frames containing information about the chemical lists. Note,
  these are not the chemical lists themselves. To access the chemicals in a given list,
  use \code{\link{get_chemicals_in_list}}.
}
\description{
Get chemical list by name batch
}
\seealso{
\code{\link{get_chemicals_in_list}}
}
