% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-APIs.R
\name{get_reported_functional_use}
\alias{get_reported_functional_use}
\title{Get Reported Functional Use}
\usage{
get_reported_functional_use(
  DTXSID = NULL,
  API_key = NULL,
  Server = "https://comptox.epa.gov/ctx-api/exposure",
  verbose = FALSE
)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}

\item{Server}{The root address for the API endpoint}

\item{verbose}{A logical indicating if some "progress report" should be given.}
}
\value{
A data.frame containing reported functional use data.
}
\description{
Get Reported Functional Use
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get reported functional use data for Caffeine
get_reported_functional_use('DTXSID0020232')
\dontshow{\}) # examplesIf}
}
