% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-APIs.R
\name{get_skin_eye_hazard}
\alias{get_skin_eye_hazard}
\title{Get skin and eye hazard}
\usage{
get_skin_eye_hazard(DTXSID = NULL, API_key = NULL)
}
\arguments{
\item{DTXSID}{The chemical identifier DTXSID}

\item{API_key}{The user-specific API key}
}
\value{
A data.frame containing skin and eye hazard data.
}
\description{
Get skin and eye hazard
}
